requirements.txt
fastapi[standard]
uvicorn
SQLAlchemy
python-dotenv